+++ lib/types/src/style.d.ts
@@ -1,0 +1,2 @@
+export declare const panelStyle: string;
+export declare const contextMenuStyle: string;