+++ lib/types/src/types/ContextMenuOption.d.ts
@@ -1,0 +1,5 @@
+export type ContextMenuOption = {
+    width?: number;
+    hideMenuWhenScroll?: boolean;
+    fixMenuWhenScroll?: boolean;
+};