+++ src/types/ContextMenuOption.ts
@@ -1,10 +1,10 @@
-export default interface ContextMenuOption {
-  /** 默认菜单宽度 */
+export type ContextMenuOption = {
+  /** default menu width */
   width?: number;
   /** 基础z-index 
   baseZIndex?: number;*/
   /** 滚动时是否关闭菜单*/
   hideMenuWhenScroll?: boolean;
-  /** 滚动时菜单是否固定(需要设置hideMenuWhenScroll=false) */
+  /** 滚动时菜单是否固定(hideMenuWhenScroll=false) */
   fixMenuWhenScroll?: boolean;
-}
+};